+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## [0.1.5]
+
+* Using fallback builder when no referenced animation is found.
+
+## [0.1.4]
+
+* Fixed staggered motions.
+
+## [0.1.3]
+
+* Fixed motions padding.
+
+## [0.1.2]
+
+* Fixed motion references.
+
 ## [0.1.1]
 
 * Improved documentation.